requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==5.0.1
autopep8==2.0.2
billiard==4.2.1
Brotli==1.0.9
celery==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.5
cryptography==39.0.0
cssmin==0.2.0
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==1.3.0
Django==4.1.5
django-allauth==0.52.0
django-appconf==1.0.5
django-auto-logout==0.5.1
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-compressor==4.3.1
django-crispy-forms==1.14.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-otp==1.1.6
django-qr-code==3.1.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
django-timezone-field==7.0
django-widget-tweaks==1.5.0
fonttools==4.39.2
geographiclib==2.0
geopy==2.4.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jsmin==3.0.1
kombu==5.4.2
lxml==4.9.2
sorl-thumbnail==12.11.0
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.6
pydyf==0.5.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyPDF3==1.0.6
pyphen==0.14.0
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.4.2
rcssmin==1.1.1
redis==5.2.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.1
segno==1.5.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.8
zopfli==0.2.2
